Revert "move i2c implementation to its own file"
This reverts commit 078c5ae6a98a5d3c18c6dfef554006fa2b18fbfb.
